move all static acls to templates
